requirements.txt
Django==1.10.7
djangorestframework==3.6.3
olefile==0.44
Pillow==4.1.1
